Trivial change to kick off AMI build.
The CircleCI environment for this repo was configured to use the wrong AWS account. As a result, the AMIs that built in the previous publish-amis branch are not in the desired AWS account and subsequent builds will fail when no example AMIs can be found. This dummy commit kicks off a build in the publish-amis branch that will build out AMIs in the correct AWS account. We can then merge the change, and re-issue the release.