Update docs to clarify role of root of repo
We've had a lot of confusion about the purpose of the code in the root of this repo. This is an attempt to improve the docs to make this more clear. Feedback welcome on how to make it even clearer.

See also https://github.com/hashicorp/terraform-aws-vault/pull/137 and https://github.com/hashicorp/terraform-aws-consul/pull/118.